+++ fabric/src/main/java/com/lx862/mtrscripting/util/FilesUtil.java
@@ -4,6 +4,8 @@ import org.apache.commons.io.FileUtils;
 import org.mtr.libraries.kotlin.text.Charsets;
 import org.mtr.mapping.holder.MinecraftClient;
 
+import javax.imageio.ImageIO;
+import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -14,23 +16,28 @@ public class FilesUtil {
     private static final Path rootMinecraftPath = MinecraftClient.getInstance().getRunDirectoryMapped().toPath();
     private static final Path dataPath = rootMinecraftPath.resolve("data").resolve("mtrscripting");
 
-    public static String read(String... paths) throws IOException {
+    public static FileEntry read(String... paths) throws IOException {
         Path p = resolvePathSafe(rootMinecraftPath, paths);
         File f = p.toFile();
         if(!f.exists()) return null;
-        return FileUtils.readFileToString(f, Charsets.UTF_8);
+        return new FileEntry(f);
     }
 
-    public static String readData(String... paths) throws IOException {
+    public static FileEntry readData(String... paths) throws IOException {
         Path p = resolvePathSafe(dataPath, paths);
         File f = p.toFile();
         if(!f.exists()) return null;
-        return FileUtils.readFileToString(f, Charsets.UTF_8);
+        return new FileEntry(f);
     }
 
     public static void saveData(String content, String... paths) throws IOException {
         Path p = resolvePathSafe(dataPath, paths);
         FileUtils.writeStringToFile(p.toFile(), content, Charsets.UTF_8);
+    }
+
+    public static void saveData(BufferedImage image, String... paths) throws IOException {
+        Path p = resolvePathSafe(dataPath, paths);
+        ImageIO.write(image, "PNG", p.toFile());
     }
 
     public static void deleteData(String... paths) throws IOException {
@@ -69,6 +76,26 @@ public class FilesUtil {
             }
             if(pPath.equals(constrainedPath)) break;
             pPath = pPath.getParent();
+        }
+    }
+
+    public static class FileEntry {
+        private final File fileReference;
+
+        public FileEntry(File fileReference) {
+            this.fileReference = fileReference;
+        }
+
+        public String asString() throws IOException {
+            return FileUtils.readFileToString(fileReference, Charsets.UTF_8);
+        }
+
+        public BufferedImage asBufferedImage() throws IOException {
+            return ImageIO.read(fileReference);
+        }
+
+        public byte[] asRawBytes() throws IOException {
+            return FileUtils.readFileToByteArray(fileReference);
         }
     }
 }